third_party/blink/web_tests/external/wpt/svg/interact/scripted/async-01-expected.txt

This is a testharness.js-based test.
[FAIL] An async script does not block the parser while downloading
  assert_equals: expected "[\\"2\\",\\"1\\"]" but got "[\\"1\\",\\"2\\"]"
Harness: the test ran to completion.
